1.2.0 Release
Added new features to enhance context-specific manipulation of the view
state.  Introduction of "Parents" collection (similar to Parent but with
all parents in high-to-low hierarchy order), Root (similar to Parent but
cuts right to the "root" of the view hierarchy) and "Family" which is an
object that has properties named for each level of the hierarchy (i.e.
item, row, column).  Added the "apply" template processing construct for
controlling the application of sub-templates.